move utility function to ExtractProcessData
